Pass the custom data install path to the compiler
The CUSTOM_DATA_INSTALL_PATH constant is checked at different
places in the code but was not actually passed to the compiler.
This resulted in searching the data files in /usr/games instead
of /usr/share/games/megaglest (in Ubuntu and Debian).

This commit also fixes setting the variable. It is still wrong
for Apple I think. But I can not test it.