Change Mac specific CMakeLists for CPack
This commit includes several fixes and some shuffling of entries to improve
readability. The major changes are Removing MG_VERSION_H_CONTENTS which
*includes* the file with the version in it in favour of VER_* variables and
removing the CPack include at the end of the file.

CPack has to be included after all variables are set and this file is sourced
before all the common options are configured.

One more fixme is added - this and other runtime issues will have to be dealt
with when i test the packaging, so far I've only got it to build.