Split up CPack generator configuration
I've taken the liberty of splitting up the CPACK_GENERATOR test so shared
configuration and win32 configuration are outside that test. The existing
grouping forced duplication of variables in the mac config file and required
the cpack generator to be unset, then set to deb, then set to nsis on win32.

These changes should help make things much more understandable.