Changes to make package build and run on osx
MegaGlest is the correct spelling, the process of building a package upper
cases the binary so I'm removing my modification via an earlier branch.

This also sets CUSTOM_DATA_INSTALL_PATH to /Applications - this is the final
step in getting MegaGlest.app to launch after building and installing.

The previously hard coded path was ok for gnumake but failed when Xcode got
involved as it has build types (release/debug/et al) which the hard coding
didn't support. I've added an if/else catch for the two build types that i've
seen while working on megaglest and it should be trivial to add any others that
may be desirable.

There is still an outstanding issue of embedding dependencies or using system
provided dependancies which I haven't sorted out.