Compatibility with current cppcheck version
Multi-threading is now disabled since it cannot be used with the
"unusedFunction" check - which we had enabled so far (and probably want to
keep enabled). Add -j and change the checks in --enable to re-enable multi
threading.

Current cppcheck also depends on a 'library' configuration file. Its location,
and that of cppcheck itself, are now configurable in this script.

skipci