display MG version number in "Real Name" field
In my IRC client, I can right click a person's nick,
and view a "Real Name" field. An MG user's Real Name field
reads "megaglest". This changes it to "MegaGlest v3.13-dev user".

I would like to #include the version info from somewhere else,
instead of initializing the variable with a constanst
string (L40).

Ideas for doing that gracefully?

Is this PR a desired change?
The reason I thought this was a good idea is
to reduce all the "What version are you using?" questions
in #megaglest-lobby.