G3D B290 script: fix importing mesh with 0 size first frame
Script use to always use the first frame as the "basis" mesh.
But this caused problems because sometimes the first mesh had a
size 0 i.e. all the verts are the same. This is fixed by adding
a step to search for a frame with the largest number of verts in
different positions. This frame is then used as the basis mesh.
Errors could still happen if the mesh doesn't have any frames
without overlapping verts.