G3D B290: remove line generating mesh code
Line was generated for meshes with duplicate verts, to prevent loss
of vertices. But this was a problem because some duplicate verts
need removing. (E.g. some duplicate verts are generated for UV
mapping. These should be removed during import.) This does mean the
g3d script can lose information on import.